posts

+++ closr-back/src/controllers/user.controller.js
@@ -36,13 +36,20 @@ export const getUsers = async (req, res) => {
   }
 };
 
-export const getUser = async (req, res) => {
+export const getMe = async (req, res) => {
   try {
     const user = await prisma.user.findUnique({
       where: { id: req.user.userId },
+      select: {
+        id: true,
+        username: true,
+        email: true,
+        role: true,
+        createdAt: true,
+      },
     });
 
-    res.json(user);
+    return res.json(user);
   } catch (error) {
     console.log(error);
   }
@@ -139,6 +146,16 @@ export const refreshAccessToken = async (req, res) => {
 export const logoutUser = async (req, res) => {
   try {
     const { refreshToken } = req.body;
+
+    const token = await prisma.refreshToken.findUnique({
+      where: { token: refreshToken },
+    });
+
+    if (!token) {
+      return res.status(404).json({
+        message: "Token no encontrado",
+      });
+    }
 
     await prisma.refreshToken.delete({
       where: { token: refreshToken },
